requirements.txt
chardet==5.2.0
faiss-cpu==1.8.0.post1
h5py==3.11.0
numpy==1.26.3
openpyxl==3.1.5
pandas==2.2.2
PyQt5==5.15.11
psutil==6.0.0
PyQt-Fluent-Widgets
sentence_transformers==3.0.1
torch
tqdm==4.66.4
transformers==4.43.3